requirements.txt
streamlit-mic-recorder
st-gsheets-connection
gspread
google-auth
